rename test classes file
